Fleshed out Blogs model, added index functionality
Blogs admin model is now easier to search and offers more relevant
information.
The index page will be housed in the Isaac_website project folder. This
will allow that folder to contain all the tight coupling of grabbing
models from the other app(s) without the risk of adding models to an Index
app and creating circular dependencies. This is probably not proper use of
the root project folder, but works as a stopgap for now.
Added startbootstrap-bare to the index as a test of the framework's
compatibility. The page appears to work as intended and can grab Post
objects from the Blogs projects and print them on the index page. The
framework will be expanded to the Blogs and other future apps.