Fleshed out Blogs model, added Index model
Blogs admin model is now easier to search and offers more relevant
information.
Index app will contain the index page for the site and will link to other
apps as needed. This allows for a more modular design of the other apps,
conceding that the Index app will suffer some tight coupling.
Added startbootstrap-bare to the Index app as a test of the framework. The
page appears to work as intended and the framework will be expanded to the
Blogs and other future apps.