Implemented dynamic startbootstrap index page
Implemented Startbootstrap-bare to pull the Post model from the Blogs app.
The index page now displays the 5 latests posts in ascending order.
As a correction from the previous commit, an Index model WAS implemented,
but was removed to avoid any possibility of a circular dependency resulting
from additional models in the Index page. The index now lives in
Isaac_website and contains the tight coupling to this specific project,
apps are still intended to be modular.