Modularized templates, moved static file root
Moved static files to the root directory. These files will not be served
by Django in production, so they were removed from the app folders for
easy linking/unlinking during deployment.

Modularized templates in the Blogs app and in the Isaac_website folder.
Header and footer (navbar at the top of the page, and css/js static files)
can now be added with {% include "[header][footer].html" %} in the content
templates. This allows the content section to be edited while maintaining
the feel and formatting of the navigation and style system across the
site.